% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{Number of trials}

\item{p}{probability of a success}
}
\value{
plot of the simulation with proportions and a table of proportions
}
\description{
This uses the sample function to produce a binomial simulation
}
\examples{
mybin()
}
