% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central Limit Theorem function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{#sample size, default}

\item{iter}{#iterations, default}

\item{b}{}
}
\value{
histogram, density curve, and theoretical normal curve
}
\description{
Central Limit Theorem function
}
