% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{A function for making a normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standard deviation}

\item{a}{}
}
\value{
a normal curve
}
\description{
A function for making a normal curve
}
